OptSmileABC class merged to OptABC
